Trait lz_str::IntoWideIter

pub trait IntoWideIter {
    type Iter: Iterator<Item = u16>;

    // Required method
    fn into_wide_iter(self) -> Self::Iter;
}

A trait to make it easier to pass arguments to functions.

Required Associated Types

type Iter: Iterator<Item = u16>

The Iterator type

Required Methods

fn into_wide_iter(self) -> Self::Iter

Convert this object into something that yields possibly invalid wide characters.